refactor: seperate HightLightText component from main
